+++ blocklets/ai-studio/src/pages/project/project-page.tsx
@@ -5,34 +5,35 @@ import {
   Button,
   ButtonProps,
   CircularProgress,
-  Paper,
   Stack,
   Tab,
   Tabs,
   Toolbar,
   Tooltip,
+  Typography,
+  tabScrollButtonClasses,
 } from '@mui/material';
 import { useLocalStorageState } from 'ahooks';
-import { bindPopper, bindTrigger, usePopupState } from 'material-ui-popup-state/hooks';
-import { useCallback, useEffect, useRef } from 'react';
+import { useCallback, useEffect, useRef, useState } from 'react';
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
 
 import { TemplateYjs } from '../../../api/src/store/projects';
 import WithAwareness from '../../components/awareness/with-awareness';
 import TemplateFormView from '../../components/template-form';
-import Popper from '../../components/template-form/popper';
-import TemplateSettings from '../../components/template-form/template-settings';
 import { useComponent } from '../../contexts/component';
 import { useIsAdmin } from '../../contexts/session';
 import ColumnsLayout, { ImperativeColumnsLayout } from './columns-layout';
 import DebugView from './debug-view';
 import FileTree, { ImperativeFileTree } from './file-tree';
 import Add from './icons/add';
-import Filtering from './icons/filtering';
+import DeveloperTools from './icons/developer-tools';
+import Empty from './icons/empty';
 import FolderAdd from './icons/folder-add';
 import PanelLeft from './icons/panel-left';
 import PanelRight from './icons/panel-right';
+import SettingView from './setting-view';
 import { useProjectState } from './state';
+import { TokenUsage } from './token-usage';
 import { isTemplate, useStore } from './yjs-state';
 
 const defaultBranch = 'main';
@@ -85,6 +86,10 @@ export default function ProjectPage() {
     const path = filename && Object.values(store.tree).find((i) => i?.endsWith(filename));
 
     if (path) navigate(path, { replace: true });
+    else {
+      const first = Object.values(store.tree)[0];
+      if (first) navigate(first, { replace: true });
+    }
   }, [gitRef, synced, location]);
 
   useEffect(() => {
@@ -112,7 +117,7 @@ export default function ProjectPage() {
   const layout = useRef<ImperativeColumnsLayout>(null);
   const fileTree = useRef<ImperativeFileTree>(null);
 
-  const settings = usePopupState({ variant: 'popper' });
+  const [currentTab, setCurrentTab] = useState('setting');
 
   return (
     <ColumnsLayout
@@ -158,7 +163,7 @@ export default function ProjectPage() {
         </Stack>
       }
       right={
-        <Stack sx={{ height: '100%' }}>
+        <Stack sx={{ height: '100%', overflow: 'auto' }}>
           <Box
             sx={{
               position: 'sticky',
@@ -168,10 +173,20 @@ export default function ProjectPage() {
               borderBottom: (theme) => `1px dashed ${theme.palette.grey[200]}`,
             }}>
             <Toolbar variant="dense" sx={{ gap: 1, px: { xs: 1 } }}>
-              <Tabs variant="fullWidth" value={0}>
-                <Tab label="Debug" />
-                <Tab label="Test" disabled />
-                <Tab label="Discuss" disabled />
+              <Tabs
+                variant="scrollable"
+                value={currentTab}
+                onChange={(_, tab) => setCurrentTab(tab)}
+                sx={{
+                  [`.${tabScrollButtonClasses.disabled}`]: {
+                    opacity: 1,
+                    color: (theme) => theme.palette.action.disabled,
+                  },
+                }}>
+                <Tab value="setting" label={t('setting')} />
+                <Tab value="debug" label={t('debug')} />
+                <Tab value="test" label={t('test')} disabled />
+                <Tab value="discuss" label={t('discuss')} disabled />
               </Tabs>
 
               <Box flex={1} />
@@ -180,11 +195,17 @@ export default function ProjectPage() {
             </Toolbar>
           </Box>
 
-          {template && <DebugView projectId={projectId} gitRef={gitRef} template={template} />}
+          {!template ? (
+            <DebugEmptyView />
+          ) : currentTab === 'setting' ? (
+            <SettingView projectId={projectId} gitRef={gitRef} template={template} />
+          ) : currentTab === 'debug' ? (
+            <DebugView projectId={projectId} gitRef={gitRef} template={template} />
+          ) : null}
         </Stack>
       }>
       {({ leftOpen, rightOpen }) => (
-        <Box sx={{ height: '100%', overflow: 'auto' }}>
+        <Stack sx={{ height: '100%', overflow: 'auto' }}>
           <Box
             sx={{
               position: 'sticky',
@@ -204,35 +225,6 @@ export default function ProjectPage() {
 
               <Box flex={1} />
 
-              {template && (
-                <>
-                  <Tooltip title={t('settings')}>
-                    <Button
-                      {...bindTrigger(settings)}
-                      color="inherit"
-                      sx={{
-                        minWidth: 0,
-                        bgcolor: settings.isOpen ? (theme) => theme.palette.action.selected : undefined,
-                      }}>
-                      <Filtering />
-                    </Button>
-                  </Tooltip>
-
-                  <Popper {...bindPopper(settings)} onClose={settings.close}>
-                    <Paper
-                      sx={{
-                        mt: -1,
-                        p: { xs: 2, md: 4 },
-                        maxWidth: 'sm',
-                        maxHeight: '90vh',
-                        overflow: 'auto',
-                      }}>
-                      <TemplateSettings projectId={projectId} gitRef={gitRef} value={template} />
-                    </Paper>
-                  </Popper>
-                </>
-              )}
-
               {!rightOpen && (
                 <PanelToggleButton
                   placement="right"
@@ -243,7 +235,7 @@ export default function ProjectPage() {
             </Toolbar>
           </Box>
 
-          <Box m={{ xs: 1 }}>
+          <Box m={{ xs: 1 }} flexGrow={1}>
             {!synced ? (
               <Box sx={{ textAlign: 'center', mt: 10 }}>
                 <CircularProgress size={32} />
@@ -254,9 +246,27 @@ export default function ProjectPage() {
               </WithAwareness>
             ) : filepath ? (
               <Alert color="error">Not Found</Alert>
-            ) : null}
+            ) : (
+              <EmptyView />
+            )}
           </Box>
-        </Box>
+
+          {template && (
+            <Box
+              sx={{
+                position: 'sticky',
+                bottom: 0,
+                bgcolor: 'background.paper',
+                zIndex: (theme) => theme.zIndex.appBar,
+                borderTop: (theme) => `1px dashed ${theme.palette.grey[200]}`,
+              }}>
+              <Toolbar variant="dense" sx={{ px: { xs: 1 } }}>
+                <TokenUsage template={template} />
+                <Box />
+              </Toolbar>
+            </Box>
+          )}
+        </Stack>
       )}
     </ColumnsLayout>
   );
@@ -275,5 +285,27 @@ function PanelToggleButton({
         {placement === 'left' ? <PanelLeft /> : <PanelRight />}
       </Button>
     </Tooltip>
+  );
+}
+
+function EmptyView() {
+  const { t } = useLocaleContext();
+
+  return (
+    <Stack color="text.disabled" alignItems="center" my={8} gap={3}>
+      <Empty sx={{ fontSize: 54, color: 'grey.300' }} />
+      <Typography variant="body2">{t('notOpenFile')}</Typography>
+    </Stack>
+  );
+}
+
+function DebugEmptyView() {
+  const { t } = useLocaleContext();
+
+  return (
+    <Stack color="text.disabled" alignItems="center" my={8} gap={3}>
+      <DeveloperTools sx={{ fontSize: 54, color: 'grey.300' }} />
+      <Typography variant="body2">{t('notOpenFile')}</Typography>
+    </Stack>
   );
 }